Corrigeer datatypes

+++ TennisWebapplication/Models/Game/Game.cs
@@ -14,7 +14,7 @@ namespace TennisWebapplication.Models
         public int Id { get; set; }
         public string GameNumber { get; set; }
         public int MemberId { get; set; }
-        public int LeagueId { get; set; }
+        public byte LeagueId { get; set; }
         public DateTime Date { get; set; }
 
         public virtual ICollection<GameResult> GameResults { get; set; }
