+++ TennisWebapplication/Models/Member/Member.cs
@@ -18,7 +18,7 @@ namespace TennisWebapplication.Models
         public string FirstName { get; set; }
         public string LastName { get; set; }
         public DateTime BirthDate { get; set; }
-        public int GenderId { get; set; }
+        public byte GenderId { get; set; }
 
         public string Address { get; set; }
         public string Number { get; set; }
