+++ TennisWebapplication/Models/MemberRole/MemberRole.cs
@@ -9,7 +9,7 @@ namespace TennisWebapplication.Models
     {
         public int Id { get; set; }
         public int MemberId { get; set; }
-        public int RoleId { get; set; }
+        public byte RoleId { get; set; }
         public DateTime StartDate { get; set; }
         public DateTime? EndDate { get; set; }
         public virtual Member MemberNavigation { get; set; }
